new bessel doc tests

+++ test/doc/bessel/airy.cpp
@@ -1,11 +1,11 @@
 // revision 0
 #include <eve/module/bessel.hpp>
-#include <iostream> 
- 
-eve::wide<float> wf([](auto i, auto c)->float{ return 2*(i-c/2);}); 
- 
-int main(){ 
-   std::cout << "<- wf = " << wf << "\n"; 
- 
+#include <iostream>
+
+eve::wide<double> wf([](auto i, auto c)->double{ return 2*(i-c/2);});
+
+int main(){
+   std::cout << "<- wf = " << wf << "\n";
+
    std::cout << "-> airy(wf)= " << eve::airy(wf) << "\n";
 }
