+++ test/doc/bessel/sph_bessel_j1.cpp
@@ -2,7 +2,7 @@
 #include <eve/module/bessel.hpp>
 #include <iostream> 
  
-eve::wide<float> wf([](auto i, auto c)->float{ return 2*(i-c/2);}); 
+eve::wide<double> wf([](auto i, auto c)->double{ return 2*(i-c/2);}); 
  
 int main(){ 
    std::cout << "<- wf = " << wf << "\n";
